+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+
+## v1.5.0 - 2025-11-24
 - When searching the local machine for available remotes, it now also searches `/rom/config/unicorn/remotes`.
 - When a file ending in `.lua` contains `require`, we now add a line to the top of the file that patches
   `package.path` to add `/lib` to the search path.